requirements.txt
selenium>=4.0.0
python-telegram-bot>=20.0
python-dotenv>=0.19.0
requests>=2.26.0
requests-oauthlib>=1.3.0 